Make JsonConverter nested
