--- src/flatcollections-array/FlatArray/FlatArrayJsonConverter/Converter.InnerItemType.cs
@@ -1,9 +1,0 @@
-namespace System;
-
-partial class FlatArrayJsonConverter<T>
-{
-    private static class InnerItemType
-    {
-        internal static readonly Type Value = typeof(T);
-    }
-}
